Added availability to skip the adaption process
L32-L36: If a GET-Parameter ('skipAdaptive') is transmitted in the HTTP-Request with value==1, then the rest of the script is ignored and the original image is sent to the browser without any adaption. (Resize, compression, etc.)

Example URL: www.yourserver.org/hqimage.jpg?skipAdaptive=1

---
I need this feature for press downloads: Regardless of the screen resolution, the visitors should be able to view/download some certain images in full quality (eg. print versions, etc.)

....ay be useful to someone else too...